requirements.txt
ddddocr
numexpr
bs4
pycryptodome
lxml